requirements.txt
torch==1.10.0
torchvision==0.11.1
numpy==1.21.5
Pillow==9.0.0
